Part out 1k line entry file in to definition headers + logically segment functionality (for future readability & extensibility)
