+++ level_data.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "game_types.h"
+#include "utilities.h"
+
+block_data * spawn_blocks_level1();
+block_data * spawn_blocks_level2();
+block_data * spawn_blocks_level3();
+block_data * spawn_blocks_level4();
