+++ rendering.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "game_types.h"
+#include "game_data.h"
+#include <splashkit.h>
+
+void draw_blocks();
+void draw_game();
+void show_title_screen();
