+++ resources.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "game_types.h"
+#include <splashkit.h>
+
+void load_resources();
+bitmap get_powerup_bitmap(powerups kind);
